Refactor Mapbox code
Change variable name to make it easier to understand, as well as related methods.

As setupMap() is quite long, it makes difficult to read it for the first time. I separate the draw and marker functions inside setupMap() into two functions to makes code cleaner and facilitate code review.

As removeLayer(), removeSource() in addRoute() and removeRoute() are the same. So I add a new function called deleteDrawing() that holds two remove function, then replace it. This new function makes easier to reuse in the future.