Refactor Facebook banner code
Change variable and function name to make it more meaningful and reasonable.

Update resizePhoto() to reuse it in scrollPhoto() by adding ‘value’ parameter to this function.

Delete uneditPhoto() and replace it with styleCoverEditBtn(type), type is ‘block’ or ‘none’. This new function makes code cleaner and easier to reuse.

As editBtn element is styled ‘block’, uploadBtn element is styled ‘none’ in three function init(), cancelUpload(), savePhoto() are the same. So I add new function called showEditBtn() that styles this two elements and I don’t need to rewrite it many times.